update
Since when you create a list of URLs, and because it must be specifically formateed correctly almost the entire actions used on it, it might as well be better to explained at the header generalizing when they refer to as the URL list.

I also added an explanation much better to sort URLs by which one that disappeared and new URLs.