improvements to the email-based site saving
I came back here because the WBGS (wayback-gsheets) frequently have issues of running extremely slow and often queuing. Not to mention that some URLs in google sheets appear to be "skipped" where there is a URL or two not having its result despite hundreds of them below it have their results being written.

I talk to the people working on it and it is due to tons of users also using the WBGS.